Update README.md
Last major change was to make the whole _python folder portable. This is an attempt to start thinking about it (custom Flame tools) as a general module structure, including it's sub-modules ...
A __init__.py file was added into _python, it only has a function to pass it's absolute parent path across the board.